ai-103-rag-eval/requirements.txt
azure-ai-evaluation>=1.16.5
azure-ai-projects>=2.0.1
azure-identity>=1.19.0
azure-search-documents>=11.6.0
azure-storage-blob>=12.22.0
openai>=1.54.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
python-multipart>=0.0.17
pypdf>=5.1.0
